refactor(frontend): split active room visibility
